requirements.txt
biopython==1.85
coverage==7.13.4
Cython==3.2.4
exceptiongroup==1.3.1
iniconfig==2.3.0
mock==5.2.0
numpy==2.4.2
packaging==26.0
pluggy==1.6.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
tomli==2.4.0
